backend/requirements.txt
fastapi==0.119.0
uvicorn==0.38.0
pydantic==2.12.3
python-dotenv==1.1.1

pymupdf==1.24.9
pdfminer.six==20250506

numpy==2.3.4
pandas==2.3.3
tiktoken==0.12.0
openai>=1.40.0

python-multipart==0.0.20